Atualiza a API
Em principal o mapper do serviço.

- Atualiza os DTO's que serão usados.
- Ajusta politica do CORS para acesso da implementaçõa ao consumo de dados
da API.

+++ Codigo/AjudAkiWeb/Service/SolicitacaoServicoService.cs
@@ -1,5 +1,5 @@
 ﻿
-using Core;
+using Core.Models;
 using Core.Service;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata;
@@ -28,7 +28,7 @@ namespace Service
             context.Add(solicitacaoServico);
             context.SaveChanges();
 
-            return solicitacaoServico.Id;
+            return (uint)solicitacaoServico.Id;
         }
 
         /// <summary>
